requirements.txt
Flask==2.2.5
numpy==1.24.3
opencv_contrib_python==4.7.0.72
opencv_python==4.10.0.84
Pillow==9.4.0
Pillow==11.0.0
tensorflow==2.18.0
tensorflow_intel==2.18.0
